requirements.txt
# DeepSecure - Python Dependencies
# Core API framework
fastapi==0.120.2
starlette==0.49.1
uvicorn==0.38.0
python-multipart==0.0.20

# Data validation
pydantic==2.12.3
pydantic_core==2.41.4
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2

# Deep learning (kept for audio model - torch)
# Note: torch is a CPU build. For GPU support, install separately via:
# pip install torch --index-url https://download.pytorch.org/whl/cu121
torch==2.6.0
torchvision==0.21.0

# Static Malware ML (EMBER-style PE feature extraction)
lief>=0.14.0
pefile>=2023.2.7
lightgbm>=4.0.0

# Image processing
pillow==10.4.0

# Audio processing
librosa==0.11.0
soundfile==0.13.1
audioread==3.1.0
soxr==1.0.0

# Machine learning (Audio classifier)
scikit-learn==1.5.2
scipy==1.13.1
joblib==1.4.2  # also used by malware model loader
threadpoolctl==3.6.0
numba==0.60.0
llvmlite==0.43.0

# Numerical computing
numpy==1.26.4

# Video processing (Temporal analysis)
opencv-python==4.10.0.84

# Utilities
requests==2.32.3
urllib3==2.2.3
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
click==8.1.8
colorama==0.4.6
h11==0.14.0
anyio==4.6.2
sniffio==1.3.1
packaging==24.2
platformdirs==4.3.6
filelock==3.16.1
fsspec==2024.10.0
networkx==3.4.2
sympy==1.13.3
mpmath==1.3.0
decorator==5.1.1
lazy_loader==0.4
msgpack==1.1.0
pooch==1.8.2
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
cffi==1.17.1

# Testing
pytest==8.3.3
httpx==0.27.2

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
python-dotenv==1.0.1
